// libs/exceptions/src/core/integration/external-service-unavailable.exception.d.ts
import { IntegrationException } from "./integration.exception.js";
export declare class ExternalServiceUnavailableException extends IntegrationException {
    constructor(serviceName: string, reason: string, data?: Record<string, unknown>);
}
//# sourceMappingURL=external-service-unavailable.exception.d.ts.map